Master Patient Index

A fabricated Master Patient Index (MPI) inspired by Gloucestershire's population to be used with functions included in SangerTools

Usage

master_patient_index

Format

A tibble with 10,000 rows and 11 variables:

PseudoNHSNumber

A Pseudonymised NHS Patient Identifier

Sex

The identifiable sex of the patient

Smoker

Health Condition Flag: 1 denotes if the patient is a smoker

Diabetes

Health Condition Flag: 1 denotes if the patient has diabetes

Dementia

Health Condition Flag: 1 denotes if the patient has dementia

Obesity

Health Condition Flag: 1 denotes if the patient is Obese

Age

Age of the patient

IMD_Decile

The decile of indices of multiple deprivation: https://www.gov.uk/government/statistics/english-indices-of-deprivation-2019

Ethnicity

The identifiable ethnicity of the patient

Locality

The region where the patient lives - sampled from Gloucestershire Clinical Commissioning Group

PrimaryCareNetwork

The network of General Practioners that the patient is registerd with - sampled from Gloucestershire Clinical Commissioning Group

Source

Generated by Asif Laldin a.laldin@nhs.net, Feb-2022

Examples

library(dplyr)
data(master_patient_index)
# Convert diabetes data to factor'
master_patient_index %>%
 glimpse()
#> Rows: 10,000
#> Columns: 10
#> $ Sex                <chr> "Female", "Female", "Male", "Female", "Female", "Fe…
#> $ Smoker             <dbl> 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, …
#> $ Diabetes           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, …
#> $ Dementia           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ Obesity            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, …
#> $ Age                <dbl> 17, 31, 34, 62, 37, 57, 25, 14, 30, 64, 45, 70, 11,…
#> $ IMD_Decile         <dbl> 7, 1, 3, 9, 4, 9, 2, 3, 3, 1, 4, 10, 3, 4, 6, 3, 4,…
#> $ Ethnicity          <chr> "White", "White", "White", "White", "White", "White…
#> $ Locality           <chr> "Gloucester City", "Stroud and Berkeley Vale", "Che…
#> $ PrimaryCareNetwork <chr> "Tewkesbury, Newent and Staunton and West Cheltenha…